big improvement: optimise the upload timing and content, slow down it

+++ HCf460/application/matlab/matlab.h
@@ -11,6 +11,6 @@
 
 extern int Fretting_detection(FIFO_DataType in_data5[4096],	double N, double pro_N, double PAD, double  offset, double rr_threshold);
 extern int quick_detection(FIFO_DataType	in_data[16384],int win_size_time,int stride_time,float time_times,float time_add,int win_size_freq,  
-										int stride_freq,  int time_accum,int xhz1, int freq_times,	double respiration_times);
+										int stride_freq,  int time_accum,int xhz1, float freq_times,	float respiration_times);
 #endif
